Move the relay macro to relay.rs
