+++ rapina-macros/src/relay.rs
@@ -1,0 +1,156 @@
+//! The `#[relay]` attribute macro.
+
+use quote::quote;
+use syn::{FnArg, ItemFn, LitStr, Pat};
+
+pub fn relay_macro_impl(
+    attr: proc_macro2::TokenStream,
+    item: proc_macro2::TokenStream,
+) -> proc_macro2::TokenStream {
+    let pattern: LitStr = syn::parse2(attr).expect("expected pattern as string literal");
+    let pattern_str = pattern.value();
+    let func: ItemFn = syn::parse2(item).expect("#[relay] must be applied to an async function");
+
+    let func_name = &func.sig.ident;
+    let func_name_str = func_name.to_string();
+
+    let is_prefix = pattern_str.ends_with('*');
+    let match_prefix_str = if is_prefix {
+        &pattern_str[..pattern_str.len() - 1]
+    } else {
+        &pattern_str
+    };
+
+    let wrapper_name = syn::Ident::new(
+        &format!("__rapina_channel_{}", func_name_str),
+        proc_macro2::Span::call_site(),
+    );
+
+    // First arg is RelayEvent (passed directly). Remaining args are extractors.
+    let args: Vec<_> = func.sig.inputs.iter().collect();
+
+    let mut extractor_extractions = Vec::new();
+    let mut call_args = vec![quote! { __rapina_event }];
+
+    for (i, arg) in args.iter().enumerate() {
+        if i == 0 {
+            // First arg is RelayEvent — passed directly, not extracted
+            continue;
+        }
+        if let FnArg::Typed(pat_type) = arg {
+            if let Pat::Ident(pat_ident) = &*pat_type.pat {
+                let arg_name = &pat_ident.ident;
+                let arg_type = &pat_type.ty;
+
+                extractor_extractions.push(quote! {
+                    let #arg_name = <#arg_type as rapina::extract::FromRequestParts>::from_request_parts(
+                        &__rapina_parts, &__rapina_params, &__rapina_state
+                    ).await?;
+                });
+
+                call_args.push(quote! { #arg_name });
+            }
+        }
+    }
+
+    quote! {
+        #func
+
+        // Generated by #[relay] — not user-facing API
+        #[doc(hidden)]
+        fn #wrapper_name(
+            __rapina_event: rapina::relay::RelayEvent,
+            __rapina_state: std::sync::Arc<rapina::state::AppState>,
+            __rapina_current_user: Option<rapina::auth::CurrentUser>,
+        ) -> std::pin::Pin<Box<dyn std::future::Future<Output = std::result::Result<(), rapina::error::Error>> + Send>> {
+            Box::pin(async move {
+                let (mut __rapina_parts, _) = rapina::http::Request::new(()).into_parts();
+                if let Some(u) = __rapina_current_user {
+                    __rapina_parts.extensions.insert(u);
+                }
+                let __rapina_params = rapina::extract::PathParams::new();
+                #(#extractor_extractions)*
+                #func_name(#(#call_args),*).await
+            })
+        }
+
+        rapina::inventory::submit! {
+            rapina::relay::ChannelDescriptor {
+                pattern: #pattern_str,
+                is_prefix: #is_prefix,
+                match_prefix: #match_prefix_str,
+                handler_name: #func_name_str,
+                handle: #wrapper_name,
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::relay_macro_impl;
+    use quote::quote;
+
+    #[test]
+    fn test_relay_macro_generates_wrapper_and_inventory() {
+        let attr = quote!("room:*");
+        let input = quote! {
+            async fn room(event: rapina::relay::RelayEvent, relay: rapina::relay::Relay) -> Result<(), rapina::error::Error> {
+                Ok(())
+            }
+        };
+
+        let output = relay_macro_impl(attr, input);
+        let output_str = output.to_string();
+
+        // Original function is preserved
+        assert!(output_str.contains("async fn room"));
+        // Wrapper function is generated
+        assert!(output_str.contains("__rapina_channel_room"));
+        // Inventory submission
+        assert!(output_str.contains("inventory :: submit !"));
+        assert!(output_str.contains("ChannelDescriptor"));
+        assert!(output_str.contains("pattern : \"room:*\""));
+        assert!(output_str.contains("is_prefix : true"));
+        assert!(output_str.contains("match_prefix : \"room:\""));
+        assert!(output_str.contains("handler_name : \"room\""));
+    }
+
+    #[test]
+    fn test_relay_macro_exact_match() {
+        let attr = quote!("chat:lobby");
+        let input = quote! {
+            async fn lobby(event: rapina::relay::RelayEvent) -> Result<(), rapina::error::Error> {
+                Ok(())
+            }
+        };
+
+        let output = relay_macro_impl(attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("is_prefix : false"));
+        assert!(output_str.contains("match_prefix : \"chat:lobby\""));
+    }
+
+    #[test]
+    fn test_relay_macro_extracts_additional_params() {
+        let attr = quote!("room:*");
+        let input = quote! {
+            async fn room(
+                event: rapina::relay::RelayEvent,
+                relay: rapina::relay::Relay,
+                log: rapina::extract::State<TestLog>,
+            ) -> Result<(), rapina::error::Error> {
+                Ok(())
+            }
+        };
+
+        let output = relay_macro_impl(attr, input);
+        let output_str = output.to_string();
+
+        // Both extractors should use FromRequestParts
+        assert!(output_str.contains("let relay ="));
+        assert!(output_str.contains("let log ="));
+        assert!(output_str.contains("FromRequestParts"));
+    }
+}
